Improve support for let statements
* Changed handling of let variables. These are now set using
xquery:eval as this allows to provide a context and the
@value-expression can be left as it is. This requires that namespaces
and previous let variables are passed to the eval-query. The XSLT now
determines the variables used in the expression and adds only those to
the eval-query.
* The Schematron that raised the issue has been used for testing as
example 3b.
* testquery.xq contained a lot of material that is not used in a
Schematron context and it has been trimmed to the content that is
relevant for an ETS derived from a Schematron schema. New name:
schematron-query-prolog.xq
* So far, namespace declarations that were specified in the Schematron
schema had to be added manually to the query prolog file. The namespace
declarations are now added automatically before the query prolog. Since
the ETF ETS model currently does not have a slot to store this
information, an ETS parameter is used.
* The generic message template did not include a proper general message
for an error in a Schematron rule as we do not know what the context is
about. A generic message has been added.
* Similar to the testquery.xq case, the standard ETS query reading the
ETS definition included material that is not relevant for ETSs derived
from Schematron. A new query file schematron-bsxets.xq that is tailored
for the Schematron context has been added.